server-java-modern: lint cleanup batch 1 (29 warnings fixed)
Resolves four classes of compiler lint warnings:

* serialVersionUID added to 5 Serializable classes
  (PriceMismatchException, GameDatabaseException,
  PatchApplicationException, GameNetworkException,
  PathfindingDebug.PathfindingDebugPanel)

* Removed 10 redundant casts:
  - GameStateUpdater (5 sites: getIcon/getIconAuthentic/usernameToHash)
  - Point.inWilderness (int / int already int)
  - FishingTrawler getLeakCount
  - Shop priceModifier (already int)
  - PcapLogger timestampMS cast (already long)
  - PluginHandler.getPlugin (Guava ClassToInstanceMap is generic)
  - AchievementSystem.setQuestStage (int already int)
  - ItemUseOnObject (object already GameObject)
  - BCrypt P_orig.clone() (array clone is covariant)

* Fixed 7 fallthrough warnings via @SuppressWarnings("fallthrough")
  (intent comments also added inline). All cases verified intentional:
  - Point.jagexPointToPoint: shorter-coord fallthrough into longer
  - Npc.updatePosition: NPC_TALK_TO falls through to common handling
  - ScriptContext.setInitialInteraction: COORDINATE/NONE chain
  - PathValidation.isPlayerBlocking: case 2 falls into default-return-false
  - SpellHandler.handleMobCast: CLAWS_OF_GUTHIX shares logic with
    SARADOMIN_STRIKE/FLAMES_OF_ZAMORAK

* hashCode added to 7 classes that overrode equals without it:
  - Player (uses usernameHash, matching equals)
  - GameObject (location/id/dir/type)
  - Item (catalogId/noted)
  - ThrowingEvent, RangeEvent, RangeEventNpc (owner)
  - ObjectRemover (object)
  No latent bug today (none used as Map keys), but a refactor
  putting any of these into a Hash collection would have silently
  broken without matching hashCode.

Verified: ant compile clean, smoke_test.sh 4/4 pass.

+++ server-java-modern/src/com/openrsc/server/content/achievement/AchievementSystem.java
@@ -237,11 +237,11 @@ public class AchievementSystem {
 	}
 
 	public void setQuestStage(final Player player, final int questID, final int stage) {
-		player.getCache().set("simpletask[" + questID + "]_stage", (int) stage);
+		player.getCache().set("simpletask[" + questID + "]_stage", stage);
 	}
 
 	public void setQuestStage(final Player player, final Achievement quest, final int stage) {
-		player.getCache().set("simpletask[" + quest.getId() + "]_stage", (int) stage);
+		player.getCache().set("simpletask[" + quest.getId() + "]_stage", stage);
 	}
 
 	public boolean playerCompletedQuest(final Player player, final Achievement quest) {
